feat: fix global scope via global module
